+++ README.md
@@ -149,7 +149,8 @@
     - [x] MenuGroup
     - [x] Menu
     - [x] MenuProduct
-    - [ ] Order
+    - [x] Order
+    - [x] OrderLineItem
     - [ ] OrderTable
     - [ ] TableGroup
