+++ src/main/java/kitchenpos/application/TableService.java
@@ -6,11 +6,11 @@ import java.util.Objects;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import kitchenpos.domain.Order;
+import kitchenpos.order.model.Order;
 import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.OrderVerifier;
+import kitchenpos.order.model.OrderVerifier;
 import kitchenpos.dto.OrderTableResponseDto;
-import kitchenpos.repository.OrderRepository;
+import kitchenpos.order.repository.OrderRepository;
 import kitchenpos.repository.OrderTableRepository;
 
 @Service
