+++ src/main/java/kitchenpos/order/application/dto/OrderCreateRequestDto.java
@@ -1,10 +1,11 @@
-package kitchenpos.dto;
+package kitchenpos.order.application.dto;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.model.Order;
+import kitchenpos.order.model.OrderStatus;
+import kitchenpos.orderline.application.dto.OrderLineCreateRequestDto;
 
 public class OrderCreateRequestDto {
     private Long orderTableId;
