+++ src/main/java/kitchenpos/order/model/OrderVerifier.java
@@ -1,8 +1,11 @@
-package kitchenpos.domain;
+package kitchenpos.order.model;
 
 import java.util.List;
 
 import org.springframework.util.CollectionUtils;
+
+import kitchenpos.domain.OrderTable;
+import kitchenpos.orderline.model.OrderLineItem;
 
 public class OrderVerifier {
     public static void validateNotCompleteOrderStatus(List<Order> orders) {
