+++ src/main/java/kitchenpos/order/repository/OrderRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.repository;
+package kitchenpos.order.repository;
 
 import java.util.List;
 
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import kitchenpos.domain.Order;
+import kitchenpos.order.model.Order;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
     @Query("SELECT o FROM Order o WHERE o.orderTableId IN (:ids)")
