+++ src/main/java/kitchenpos/orderline/application/dto/OrderLineItemResponseDto.java
@@ -1,9 +1,9 @@
-package kitchenpos.dto;
+package kitchenpos.orderline.application.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.orderline.model.OrderLineItem;
 
 public class OrderLineItemResponseDto {
     private final Long seq;
