+++ src/main/java/kitchenpos/orderline/repository/OrderLineItemRepository.java
@@ -1,10 +1,10 @@
-package kitchenpos.repository;
+package kitchenpos.orderline.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.orderline.model.OrderLineItem;
 
 public interface OrderLineItemRepository extends JpaRepository<OrderLineItem, Long> {
     List<OrderLineItem> findAllByOrderId(Long orderId);
